Update state tracking to include configuration values.
Configuration values will be read at process startup and then persisted the first time any config values is read.  This means that the values will not be easily influenced by parameters changing across restarts of the framework.

Zk configuration is now read from environment variables so that it can be easily configured.