Add new config flag CASSANDRA_DATA_DIRECTORY.
This is a temporally flag that will ultimately be replaced by persistent
volumes when MESOS-1554 is released.

The default data directory is set to be in the mesos sandbox following
the pattern set my Cassandra itself as of 2.1.

If the default directory is used a warning message is logged at process
start time.