Update offer evaluation logic to take resources into consideration before launching an executor.
There is a scenario in the current implementation of task launching where it can become impossible to actually launch a server task. To try and mitigate this issue when an offer is being evaluated and it is determined a new node should be started the resources in the offer will be evaluated to ensure that there are enough resources to launch the executor and all of it's tasks, if there aren't enough resources the offer will be declined.

Long term this will change to leverage the "acceptOffer" functionality offered by mesos where a framework will be able to claim resources without necessarily running a task. This will allow us to claim all the resources we need on a host and then start the bootstrapping process without having to worry about the resources being lost between the stages of task launching.