Add Health check endpoints.
Three new health check endpoints are now exposed in the REST Api.

First, there is a simple ping endpoint (/health/process) that can be used to test if the http server is up and running.

Second, there is an endpoint (/health/cluster) that attempts to return either a 200 if healthy or a 500 if unhealthy. This endpoint is specifically tailored to be called from marathon with configured maxConsecutiveFailures value of 0.  This makes it a purely informational health check rather than something marathon should react to.

Third, there is an endpoint (/health/cluster/report) that will expose the full health check report that is used to evaluate if the cluster is healthy so that a 200 or 500 can be returned from /health/cluster.  The structure of the report is such that it lists the value(s) we would expect to see from a healthy cluster for a particular evaluation point and the value(s) we actually found when analyzing the info we have about the cluster.

Currently implemented checks are:
* node count
* seed count
* all nodes "healthy" (property from executor health check)
* all nodes "operating normally"
* all nodes have a health check within the past 5 minutes