Update CassandraServerRunTask to include healthCheckIntervalSeconds.
There is currently a scenario where the framework scheduler can be starved of resource offers and is unable to send framework messages to tell nodes to run a health check.  To address this potential scenario the health check interval is now defined as a property of the CassandraServerRunTask.  When the task is received by the executor it will start a background task to perform a health check every healthCheckIntervalSeconds and send a framework message to the scheduler.